requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
pydantic==2.10.5
pydantic-settings==2.7.1

psycopg2-binary==2.9.10

celery==5.4.0
redis==5.2.1

httpx==0.28.1
beautifulsoup4==4.12.3

chromadb==0.5.23
dj-database-url==2.3.0
gunicorn==23.0.0
